move up method of basetypecheker into compositechecker
